+++ models/__tests__/poisson.test.ts
@@ -1,7 +1,0 @@
-import Poisson from '../poisson'
-
-describe('models/poisson', () => {
-  test('description', () => {
-    expect(Poisson).toBeDefined()
-  })
-})